requirements.txt
livekit-agents
livekit-agents[google]~=1.2
livekit-plugins-noise-cancellation~=0.2
python-dotenv
livekit-plugins-deepgram